feat: parse enum struct variants
