+++ crates/tuono/src/typescript/parser/parse_struct.rs
@@ -1,25 +1,9 @@
 use crate::typescript::parser::utils::{
-    RenameSerdeOptions, parse_generics_to_typescript_string, parse_serde_attribute,
+    RenameSerdeOptions, get_field_name, parse_generics_to_typescript_string, parse_serde_attribute,
+    rust_to_typescript_type,
 };
-use crate::typescript::utils::type_to_typescript;
-
-use syn::{self, GenericArgument, PathArguments};
 
 use super::utils::should_skip_element;
-
-fn get_field_name(field: &syn::Field) -> String {
-    let field_name: String = parse_serde_attribute(&field.attrs, "rename");
-
-    if !field_name.is_empty() {
-        return field_name;
-    }
-
-    if let Some(field) = field.ident.as_ref() {
-        field.to_string()
-    } else {
-        String::from("unknown")
-    }
-}
 
 /// Parse a rust struct and returns a tuple of the struct name and the
 /// struct compiled to a typescript interface
@@ -38,78 +22,7 @@ pub fn parse_struct(element: &syn::ItemStruct) -> (String, String) {
 
         let field_name = get_field_name(field);
 
-        let field_type = match &field.ty {
-            syn::Type::Path(type_path) => {
-                if let Some(last_segment) = type_path.path.segments.last() {
-                    let outer_type = last_segment.ident.to_string();
-                    if let PathArguments::AngleBracketed(args) = &last_segment.arguments {
-                        let inner_types: Vec<String> = args
-                            .args
-                            .iter()
-                            .filter_map(|arg| {
-                                if let GenericArgument::Type(inner_type) = arg {
-                                    match inner_type {
-                                        syn::Type::Path(inner_type_path) => {
-                                            Some(inner_type_path.path.segments[0].ident.to_string())
-                                        }
-                                        syn::Type::Reference(reference) => {
-                                            if let syn::Type::Path(inner_type_path) =
-                                                &*reference.elem
-                                            {
-                                                Some(
-                                                    inner_type_path.path.segments[0]
-                                                        .ident
-                                                        .to_string(),
-                                                )
-                                            } else {
-                                                Some("unknown".to_string())
-                                            }
-                                        }
-                                        _ => Some("unknown".to_string()),
-                                    }
-                                } else {
-                                    None
-                                }
-                            })
-                            .collect();
-
-                        match outer_type.as_str() {
-                            "Option" => {
-                                format!("{} | null", type_to_typescript(&inner_types[0]))
-                            }
-                            "Vec" => {
-                                format!("{}[]", type_to_typescript(&inner_types[0]))
-                            }
-                            "HashMap" | "BTreeMap" => {
-                                format!(
-                                    "Record<{}, {}>",
-                                    type_to_typescript(&inner_types[0]),
-                                    type_to_typescript(&inner_types[1])
-                                )
-                            }
-                            _ => "unknown".to_string(),
-                        }
-                    } else {
-                        type_to_typescript(&outer_type).to_string()
-                    }
-                } else {
-                    "unknown".to_string()
-                }
-            }
-            syn::Type::Reference(reference) => {
-                // Ignore lifetimes and treat references as their base type
-                if let syn::Type::Path(type_path) = &*reference.elem {
-                    if let Some(base_type) = type_path.path.segments.last() {
-                        type_to_typescript(&base_type.ident.to_string()).to_string()
-                    } else {
-                        "unknown".to_string()
-                    }
-                } else {
-                    "unknown".to_string()
-                }
-            }
-            _ => "unknown".to_string(),
-        };
+        let field_type = rust_to_typescript_type(&field.ty);
 
         let field_name = rename_option.transform(field_name);
